+++ HISTORY.rst
@@ -6,6 +6,7 @@ Release History
 0.6.9 (?)
 +++++++++
 - Added support for \n and \r characters in upload headers.
+- Added support for reading filenames from stdin when using the `ia delete` command.
 
 0.6.8 2014-07-11 
 +++++++++